requirements.txt
# Core RAG & LangChain
langchain>=0.1.16
openai>=0.27.0

# Embeddings
sentence-transformers>=2.7.0

# Vector stores
faiss-cpu>=1.8.0        # or faiss-gpu if you have CUDA
chromadb>=0.4.24
qdrant-client>=1.8.0
langchain-qdrant>=0.1.4

# PDF loading & parsing
pymupdf>=1.22.0         # for PyMuPDFLoader
pypdf>=3.8.0            # for PyPDFLoader

# Sparse retrieval (BM25)
rank-bm25>=0.2.2

# HTTP clients (for REST embedding, etc.)
requests>=2.28.0

# Optional (for CLI/env)
python-dotenv>=1.0.0

# Ollama LLM & Embeddings integration
ollama>=0.1.3                     # Ollama Python client :contentReference[oaicite:0]{index=0}
langchain-ollama>=0.1.0           # LangChain ↔ Ollama integration :contentReference[oaicite:1]{index=1}
